--- client/Login/public/dashboard.html
@@ -1,17 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>Incoming Patients Dashboard</title>
-    <link rel="stylesheet" href="styles.css">
-</head>
-<body>
-    <div class="container-dta">
-        <h1>Incoming Patients</h1>
-        <div id="patient-list"></div>
-        <button id="logout-button" class="submit-btn">Logout</button>
-    </div>
-    <script src="dashboard.js"></script>
-</body>
-</html>
