--- client/Login/public/login-styles.css
@@ -1,210 +1,0 @@
-/* General page layout */
-body {
-  font-family: 'Segoe UI', Tahoma, Geneva, Verdana, sans-serif;
-  background-color: #1c1c1c; /* Dark background */
-  margin: 0;
-  padding: 0;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  height: 100vh;
-}
-
-.container {
-  width: 100%;
-  max-width: 400px; /* Smaller width for login form */
-  padding: 30px;
-  background-color: #333; /* Dark background for the container */
-  border-radius: 12px; /* Rounded corners */
-  box-shadow: 0px 6px 24px rgba(0, 0, 0, 0.3); /* Darker shadow */
-}
-
-h1 {
-  text-align: center;
-  color: #fff; /* White text for contrast */
-  margin-bottom: 25px;
-  font-size: 28px;
-  font-weight: 600;
-}
-
-/* Input field styling */
-.input-group {
-  margin-bottom: 20px;
-}
-
-.label {
-  display: block;
-  font-size: 14px;
-  margin-bottom: 8px;
-  color: #ddd; /* Light label text */
-}
-
-input {
-  width: 100%;
-  padding: 12px 15px;
-  background-color: #444; /* Dark input background */
-  border: 1px solid #555; /* Darker border */
-  border-radius: 6px;
-  color: white;
-  font-size: 16px;
-  transition: border-color 0.3s;
-  box-sizing: border-box; /* Prevents overflow */
-}
-
-input::placeholder {
-  color: #bbb; /* Lighter color for placeholder text */
-}
-
-input:focus {
-  border-color: #007bff; /* Blue border on focus */
-  outline: none; /* Remove outline */
-}
-
-/* Submit button */
-button.submit-btn {
-  width: 100%;
-  background-color: #28a745; /* Green submit button */
-  color: white;
-  padding: 14px 0;
-  border: none;
-  border-radius: 8px;
-  font-size: 18px;
-  cursor: pointer;
-  transition: background-color 0.3s ease, box-shadow 0.3s ease;
-  margin-top: 10px;
-}
-
-button.submit-btn:hover {
-  background-color: #218838; /* Darker green on hover */
-  box-shadow: 0px 4px 12px rgba(0, 0, 0, 0.5); /* Hover shadow */
-}
-
-/* Error or info message */
-.message {
-  text-align: center;
-  color: #ff4d4d; /* Red error message */
-  margin-bottom: 15px;
-}
-
-/* Add spacing between form and login link */
-.form-footer {
-  text-align: center;
-  color: #bbb;
-}
-
-.form-footer a {
-  color: #007bff;
-  text-decoration: none;
-}
-
-.form-footer a:hover {
-  text-decoration: underline;
-}
-
-.container-dta {
-  width: 100%;
-  max-width: 800px; /* Smaller width for login form */
-  padding: 30px;
-  background-color: #333; /* Dark background for the container */
-  border-radius: 12px; /* Rounded corners */
-  box-shadow: 0px 6px 24px rgba(0, 0, 0, 0.3); /* Darker shadow */
-}
-
-/* General page layout */
-body {
-  font-family: 'Segoe UI', Tahoma, Geneva, Verdana, sans-serif;
-  background-color: #1c1c1c; /* Dark background */
-  margin: 0;
-  padding: 0;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  height: 100vh; /* Full viewport height */
-}
-
-.container-dta {
-  width: 100%;
-  max-width: 800px;
-  padding: 30px;
-  background-color: #333; /* Dark background for the container */
-  border-radius: 12px; /* Rounded corners */
-  box-shadow: 0px 6px 24px rgba(0, 0, 0, 0.3); /* Darker shadow */
-  box-sizing: border-box; /* Ensure padding doesn't affect width */
-  display: flex;
-  flex-direction: column; /* Column layout */
-  justify-content: space-between; /* Space between content and button */
-  height: 80vh; /* Fixed height, won't scroll */
-}
-
-h1 {
-  text-align: center;
-  color: #fff; /* White text for contrast */
-  margin-bottom: 25px;
-  font-size: 28px;
-  font-weight: 600;
-}
-
-/* Patient list */
-#patient-list {
-  overflow-y: auto; /* Make the patient list scrollable */
-  flex-grow: 1; /* Take up remaining space */
-  padding-right: 10px; /* Add padding to avoid content touching the edge */
-  margin-bottom: 20px; /* Add space between patient list and button */
-  max-height: calc(100% - 60px); /* Ensure patient list takes up space and allows the button to fit */
-}
-
-/* Each patient div styling */
-.patient-div {
-  margin-bottom: 15px;
-  padding: 20px;
-  background-color: #444; /* Dark background for each patient block */
-  border-radius: 8px;
-  color: white;
-  box-shadow: 0px 4px 12px rgba(0, 0, 0, 0.2); /* Subtle shadow for depth */
-  display: flex;
-  flex-direction: column;
-}
-
-.patient-div h3 {
-  margin: 0 0 8px 0;
-  font-size: 18px;
-  color: #ffcc00; /* Highlight hospital name in yellow */
-}
-
-.patient-div p {
-  margin: 4px 0;
-  font-size: 14px;
-  color: #ccc; /* Light gray for less important text */
-}
-
-/* Logout button */
-button#logout-button {
-  background-color: #28a745; /* Green submit button */
-  color: white;
-  padding: 14px 0;
-  border: none;
-  border-radius: 8px;
-  font-size: 18px;
-  cursor: pointer;
-  transition: background-color 0.3s ease, box-shadow 0.3s ease;
-  margin-top: 10px;
-  width: 100%;
-}
-
-button#logout-button:hover {
-  background-color: #218838; /* Darker green on hover */
-  box-shadow: 0px 4px 12px rgba(0, 0, 0, 0.5); /* Hover shadow */
-}
-/* CSS for severity colors */
-.hospital-low {
-  color: #4caf50; /* Green for low severity */
-}
-
-.hospital-medium {
-  color: #ff9800; /* Orange for medium severity */
-}
-
-.hospital-high {
-  color: #f44336; /* Red for high severity */
-}
-
