--- client/Login/public/login.html
@@ -1,27 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>Hospital Login</title>
-    <link rel="stylesheet" href="login-styles.css">
-</head>
-<body>
-    <div class="container">
-        <h1>Hospital Login</h1>
-        <form id="login-form">
-            <div class="form-group">
-                <label for="hospital-name" class="label">Hospital Name</label>
-                <input type="text" id="hospital-name" name="hospital-name" placeholder="Enter hospital name" required>
-            </div>
-            <div class="form-group">
-                <label for="password" class="label">Password</label>
-                <input type="password" id="password" name="password" placeholder="Enter password" required>
-            </div>
-            <button type="submit" class="submit-btn">Login</button>
-        </form>
-        <div id="error-message" class="message"></div>
-    </div>
-    <script src="login.js"></script>
-</body>
-</html>
