+++ client/Patients/static/login.js
@@ -1,34 +1,30 @@
-document.getElementById('login-form').addEventListener('submit', async function(event) {
-    event.preventDefault();
+document.getElementById('login-form').addEventListener('submit', async (e) => {
+    e.preventDefault();
     
-    const formData = {
-        hospitalName: document.getElementById('hospitalName').value,
-        password: document.getElementById('password').value
-    };
+    const hospitalName = document.getElementById('hospital-name').value;
+    const password = document.getElementById('password').value;
+    const errorMessage = document.getElementById('error-message');
 
     try {
-        console.log('Attempting login with:', formData.hospitalName);
-        
         const response = await fetch('/api/login', {
             method: 'POST',
             headers: {
-                'Content-Type': 'application/json'
+                'Content-Type': 'application/json',
             },
-            body: JSON.stringify(formData)
+            body: JSON.stringify({ hospitalName, password }),
         });
 
         const data = await response.json();
-        
+
         if (response.ok) {
-            console.log('Login successful');
             localStorage.setItem('token', data.token);
-            window.location.href = '/dashboard';
+            window.location.href = '/dashboard';  // Redirect to patients page
         } else {
-            console.error('Login failed:', data.message);
-            alert(data.message || 'Login failed');
+            errorMessage.textContent = data.message;
+            errorMessage.style.display = 'block';
         }
     } catch (error) {
-        console.error('Login error details:', error);
-        alert('Login error: ' + error.message);
+        errorMessage.textContent = 'An error occurred. Please try again.';
+        errorMessage.style.display = 'block';
     }
 });
