+++ client/Patients/templates/ambulance.html
@@ -77,7 +77,7 @@
     <script src="https://unpkg.com/leaflet/dist/leaflet.js"></script>
     <script src="https://unpkg.com/leaflet-routing-machine/dist/leaflet-routing-machine.js"></script>
     <script src="https://cdnjs.cloudflare.com/ajax/libs/Turf.js/6.5.0/turf.min.js"></script>
-    <script type="module" src="{{ url_for('static', filename='main.js') }}"></script>
+    <script type="module" src="../static/main.js"></script>
 
   </body>
 </html>
